Create a QC-metrics file

Extract QC metrics and outlier information from IDAT files. Returns summarized QC metrics, sample and probe outliers, relatedness info and DNA-methylation-based predictors (PMSs) of various traits/exposures in a QCmetrics object, which can be queried interactively using the QCapp method.

getQCmetrics(
  samplesheet,
  name = NULL,
  experimenter = NULL,
  chunk_size = 40,
  geno = NULL,
  snplinker = NULL,
  logfile = NULL,
  verbose = TRUE,
  probe_qc_method = c("post_sampleqc", "pre_sampleqc"),
  heatmap_variables = NULL,
  heatmap_variables_fct = NULL,
  sample_qc_thresholds = NULL,
  probe_qc_thresholds = NULL,
  detp_threshold = 1e-16,
  beadnr_threshold = 3,
  relatedness_thresholds = list(IBS_mean = 1.5, IBS_var = 0.35),
  custom_outliers = NULL,
  epidish_method = c("RPC", "CBS", "CP"),
  tissue = c("blood", "breast", "epithelium"),
  calculate_PCs = FALSE,
  n_probes_PCA = NULL,
  n_estimate_PCA = chunk_size,
  nrPCs = 10,
  save_all_relationships = NULL,
  path_failed_measurements = NULL,
  cores = 1,
  BPPARAM = NULL,
  force = FALSE
)

Arguments

samplesheet

Samplesheet. Should contain the following columns: Sample_Name','Sex' (M/F) and 'Basename'.

name

Optionally, provide a name for the experiment (will be included in metadata).

experimenter

Optionally, name of the experiment (will be included in metadata).

chunk_size

Size of chunks, default = 40. Larger chunks will be a little bit faster but will require more RAM.

geno

An optional matrix with genotypes (rows = snps, columns = samples) to check for concordance between DNAm and genotype data. See the tutoriols for more information.

snplinker

Two-column data frame that links DNAm probes to genotype probes (first column = DNAm probes, second column = genotype IDs). If not provided the 65 designated SNP-probes will be used.

logfile

Filepath to write logfile to, by default no logfile is written (`logfile = NULL`).

verbose

If `TRUE` (the default), information is printed in the console.

probe_qc_method

Should probe QC be performed after sample QC ('post_sampleqc') or prior to sample QC ('pre_sampleqc')?. If 'post_sampleqc', sample outliers will be removed before performing site QC.

heatmap_variables

Character vector specifying additional variables to test for association with PCs (default = `NULL`).

heatmap_variables_fct

Relevant if the `heatmap_variables` argument is not `NULL`. Which of the variables specified in `heatmap_variables` should be treated as factors?

sample_qc_thresholds

List of sample QC thresholds, if `NULL` (the default), default thresholds will be used. Note that custom thresholds can be specified for a subset of metrics, default thresholds will then be used for metrics that are not specified.

probe_qc_thresholds

List of probe QC thresholds, if `NULL` (the default), default thresholds will be used.

detp_threshold

detection P-value threshold (default = 1e-16).

beadnr_threshold

beadnr threshold (default = 3).

relatedness_thresholds

Initial relatedness thresholds, can be modified in the QCapp.

custom_outliers

Optional additional column name(s) in the samplesheet that that indicate that samples should be treated as outliers. The indicated columns should be of class `logical`, (`TRUE` = outlier).

epidish_method

Which method to use in the EpiDish epidish cellcount prediction (RPC, CBS, or CP). Default = 'RPC'.

tissue

Tissue-type for cellcount-estimation. Currently 'blood', 'breast' and 'epithelium' are implemented, following the epidish function.

calculate_PCs

Should PCA be performed? (TRUE/FALSE). Default = `FALSE`. For large sample sizes RAM could become a bottleneck. To limit RAM usage (besides setting this parameter to `FALSE` to skip PCA), the `n_probes_PCA` argument can be used to limit PCA to a subset of most variable probes.

n_probes_PCA

Number of most variable probes to maintain for PCA, defaults to 'NULL' (all probes are used).

n_estimate_PCA

Relevant if `n_probes_PCA` is not `NULL`. This argument specifies the number of samples used to estimate the most variable probes (defaults to 'chunk_size').

nrPCs

nr of PCs to retain. Default = 10

save_all_relationships

Save all relatedness data (IBS)? Relatedness output contains N^2 rows, and thus it will rapidly increase in size with large sample sizes. If `FALSE`, relations defined as unrelated (based on the relatedness thresholds) will not be saved. If `NULL` (the default), the parameter will be set to `FALSE` for sample sizes >= 1000, and to `TRUE` otherwise.

path_failed_measurements

Filepath (.rds extension) for saving a matrix that indicates for each measurement whether it failed based on the detP and beadNr cutoffs. Defaults to `NULL`, in which case this matrix is not saved. This option is useful for the mask_measurements function, which sets failed measurements in a beta-matrix to missing.

cores

Number of cores to use, defaults to 1. The `BPPARAM` argument controls parallelization (see: BiocParallel).

BPPARAM

Relevant if the `cores` argument > 1. Specifies the parallel backend to use (see BiocParallel), defaults to MulticoreParam

force

See the minfi documentation: Should reading different size IDAT files be forced? Defaults to FALSE.

Value

A QCmetrics object.

Details

Although many parameters can be tweaked, only the `samplesheet` parameter is mandatory. Note that the samplesheet must contain the following columns: Sample_Name','Sex' (M/F) and 'Basename'. The 'Basename' column should contain paths to the IDAT files. Optionally, 'Age', 'Smoking', 'BMI' and 'Alcohol' columns can be included. If these are provided provided, these can be compared to DNA-methylation-based predictors (PMSs) of the respective variables in the QCapp. All other included columns are kept and can be used within the QCapp to color the variables by those columns

Examples

## Load example data from minfiData package (6 samples)
library(minfiData)
#> Loading required package: minfi
#> Loading required package: GenomicRanges
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> Loading required package: DelayedArray
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘matrixStats’
#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: BiocParallel
#> 
#> Attaching package: ‘DelayedArray’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges
#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> Loading required package: bumphunter
#> Loading required package: foreach
#> Loading required package: iterators
#> Loading required package: locfit
#> locfit 1.5-9.4 	 2020-03-24
#> Loading required package: IlluminaHumanMethylation450kmanifest
#> Loading required package: IlluminaHumanMethylation450kanno.ilmn12.hg19
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:minfi’:
#> 
#>     combine
#> The following objects are masked from ‘package:Biostrings’:
#> 
#>     collapse, intersect, setdiff, setequal, union
#> The following object is masked from ‘package:XVector’:
#> 
#>     slice
#> The following object is masked from ‘package:matrixStats’:
#> 
#>     count
#> The following object is masked from ‘package:Biobase’:
#> 
#>     combine
#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union
#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect
#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union
#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union
#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
baseDir <- system.file("extdata", package="minfiData")
samplesheet <- read.metharray.sheet(baseDir)
#> [read.metharray.sheet] Found the following CSV files:
#> [1] "/hpc/hers_en/phop/software/R-3.6.3/libs/minfiData/extdata/SampleSheet.csv"
# Make sure that the samplesheet contains a 'Sample_Name'
# column with the sample IDs and a 'Sex' column!
samplesheet <- samplesheet %>%
                   dplyr::mutate(Sample_Name =  paste(samplesheet$Slide,
                                 samplesheet$Array, sep = "_")) %>%
                   dplyr::rename(Sex = sex)

## Get QC metrics
metrics <- getQCmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))
#> Creating a QCmetrics object
#> Analysis started at: 2022-09-27 08:55:49
#> Array: 450k
#> -----------------------------
#> Extracting QCmetrics in chunks of 40 samples..
#> No methods found in package ‘RSQLite’ for request: ‘dbListFields’ when loading ‘lumi’
#> Using 144/144 probes for alcohol prediction
#> Using 187/187 probes for smoking prediction
#> Using 435/435 probes for BMI prediction
#> Done!
#> Merging the chunks..
#> Performing IBS..
#> 10 out of 65 probes could be accurately called
#> Calculating relatedness using the DNAm-inferred SNPs
#> Performing array-wide PCA..
#> Performing control probe PCA..
#> Performing association tests between array-wide PCs and technical/biological variables..
#> The following variables have zero variance, and will therefore be excluded from the association tests:
#> 	CD8T
#> 	Neutro
#> Performing association tests between control probe PCs and technical/biological variables..
#> The following variables have zero variance, and will therefore be excluded from the association tests:
#> 	CD8T
#> 	Neutro
#> 
#> Merging the metrics..
#> 
#> Number of outliers using default thresholds:
#> MU: 0 outliers (0%)
#> RG_ratio: 0 outliers (0%)
#> GR_ratio: 0 outliers (0%)
#> OP: 0 outliers (0%)
#> HC: 0 outliers (0%)
#> bscon: 0 outliers (0%)
#> detectionP: 2 outliers (33.33%) [WARNING] High percentage of outliers!
#> beadNr: 0 outliers (0%)
#> sex: 0 outliers (0%)
#> Total: 2 outliers (33.33%) [WARNING] High percentage of outliers!
#> Done!
#> Analysis finished at: 2022-09-27 08:56:49
#> 

# Add some custom thresholds:
# Note that you can specify only the thresholds that you want to change,
# for the others the default values will be used
 thresholds <- list(MU = 2000, OP = 13, RG_ratio = 0.6)
 metrics <- getQCmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  sample_qc_thresholds = thresholds,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))
#> Creating a QCmetrics object
#> Analysis started at: 2022-09-27 08:56:49
#> Array: 450k
#> -----------------------------
#> Not all sample QC thresholds provided, defaults will be used for the following thresholds:
#> GR_ratio
#> HC
#> bscon
#> detP
#> beadNr
#> IBS_mean
#> IBS_var
#> XY_diff
#> custom_outliers
#> Extracting QCmetrics in chunks of 40 samples..
#> Using 144/144 probes for alcohol prediction
#> Using 187/187 probes for smoking prediction
#> Using 435/435 probes for BMI prediction
#> Done!
#> Merging the chunks..
#> Performing IBS..
#> 10 out of 65 probes could be accurately called
#> Calculating relatedness using the DNAm-inferred SNPs
#> Performing array-wide PCA..
#> Performing control probe PCA..
#> Performing association tests between array-wide PCs and technical/biological variables..
#> The following variables have zero variance, and will therefore be excluded from the association tests:
#> 	CD8T
#> 	Neutro
#> Performing association tests between control probe PCs and technical/biological variables..
#> The following variables have zero variance, and will therefore be excluded from the association tests:
#> 	CD8T
#> 	Neutro
#> 
#> Merging the metrics..
#> 
#> Number of outliers using user-defined thresholds:
#> MU: 0 outliers (0%)
#> RG_ratio: 0 outliers (0%)
#> GR_ratio: 0 outliers (0%)
#> OP: 0 outliers (0%)
#> HC: 0 outliers (0%)
#> bscon: 0 outliers (0%)
#> detectionP: 2 outliers (33.33%) [WARNING] High percentage of outliers!
#> beadNr: 0 outliers (0%)
#> sex: 0 outliers (0%)
#> Total: 2 outliers (33.33%) [WARNING] High percentage of outliers!
#> Done!
#> Analysis finished at: 2022-09-27 08:57:27
#>