Save QC plots contained in a QCmetrics object in high quality.

Save QC plots contained in a QCmetrics object in high quality.

savePlots(
  qcmetrics,
  outdir,
  name,
  background = NULL,
  background_alpha = 0.3,
  width = 7,
  height = 5,
  unit = "in",
  filetype = "png",
  dpi = 300,
  colorby = "None",
  colorbyQCmetric = "None",
  show_outliers = TRUE,
  show_which_outliers = c("all", "metric"),
  selected_samples = c(),
  probes_plot_cutoff = 0.01,
  text_size = 13,
  point_size = 2.5,
  point_size_outlier = point_size * 1.5,
  outlier_col = "red",
  alpha = 0.8,
  theme = theme_classic(),
  colorscheme = NULL
)

Arguments

qcmetrics

QCmetrics object

outdir

path to the directory where the plots should be save (will be created if it does not exist)

name

filename

background

A QCmetrics or QCmetricsBG object that will be used as background. Defaults to `NULL`.

background_alpha

transparency, defaults to 0.3.

width

Plot size in units ("in", "cm", or "mm"). Defaults to 7.

height

Plot size in units ("in", "cm", or "mm"). Defaults to 5.

unit

"in", "cm", or "mm". Defaults to "in"

filetype

Device to use ("png", "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf). Defaults to "png".

dpi

Plot resolution. Defaults to 300.

colorby

Column to color the plot by. Defaults to "None".

colorbyQCmetric

Color by outlier type (MU_outlier etc.)

show_outliers

Show outliers? (TRUE/FALSE). Based on outliers as defined in the QCmetrics object. Defaults to `TRUE`.

show_which_outliers

"all" or "metric". The former will highlight all outliers (based on all metrics), the latter will only the outliers based on the metric plotted.

selected_samples

Highlight user-provided vector of Sample Names.

probes_plot_cutoff

Limit the number of probes plotted in the probe QC plots. Defaults to 0.01 (only showing probes with percentage > 0.01).

text_size

Text size, defaults to 13.

point_size

Point size, defaults to 2.5.

point_size_outlier

Point size for highlighted outliers, defaults to

outlier_col

Color used for highlighting outliers, defaults to "red".

alpha

transparency, defaults to 0.8.

theme

ggplot theme to use, defaults to "theme_classic()"

colorscheme

which colorscheme to use, defaults to ggplot2::scale_color_viridis_c()/ggplot2::scale_color_viridis_d()

Examples

## Load example data from minfiData package (6 samples)
library(minfiData)
library(dplyr)
baseDir <- system.file("extdata", package="minfiData")
samplesheet <- read.metharray.sheet(baseDir)
#> [read.metharray.sheet] Found the following CSV files:
#> [1] "/hpc/hers_en/phop/software/R-3.6.3/libs/minfiData/extdata/SampleSheet.csv"
# Make sure that the samplesheet contains a 'Sample_Name'
# column with the sample IDs and a 'Sex' column!
samplesheet <- samplesheet %>%
                   dplyr::mutate(Sample_Name =  paste(samplesheet$Slide,
                                 samplesheet$Array, sep = "_")) %>%
                   dplyr::rename(Sex = sex)

## Get QC metrics
metrics <- getQCmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))
#> Creating a QCmetrics object
#> Analysis started at: 2022-09-27 08:57:29
#> Array: 450k
#> -----------------------------
#> Extracting QCmetrics in chunks of 40 samples..
#> Using 144/144 probes for alcohol prediction
#> Using 187/187 probes for smoking prediction
#> Using 435/435 probes for BMI prediction
#> Done!
#> Merging the chunks..
#> Performing IBS..
#> 10 out of 65 probes could be accurately called
#> Calculating relatedness using the DNAm-inferred SNPs
#> Performing array-wide PCA..
#> Performing control probe PCA..
#> Performing association tests between array-wide PCs and technical/biological variables..
#> The following variables have zero variance, and will therefore be excluded from the association tests:
#> 	CD8T
#> 	Neutro
#> Performing association tests between control probe PCs and technical/biological variables..
#> The following variables have zero variance, and will therefore be excluded from the association tests:
#> 	CD8T
#> 	Neutro
#> 
#> Merging the metrics..
#> 
#> Number of outliers using default thresholds:
#> MU: 0 outliers (0%)
#> RG_ratio: 0 outliers (0%)
#> GR_ratio: 0 outliers (0%)
#> OP: 0 outliers (0%)
#> HC: 0 outliers (0%)
#> bscon: 0 outliers (0%)
#> detectionP: 2 outliers (33.33%) [WARNING] High percentage of outliers!
#> beadNr: 0 outliers (0%)
#> sex: 0 outliers (0%)
#> Total: 2 outliers (33.33%) [WARNING] High percentage of outliers!
#> Done!
#> Analysis finished at: 2022-09-27 08:58:06
#> 

## Save plots
if (FALSE) {
savePlots(metrics,
          outdir = "./",
          name = "example_minfidata")
}

## color by 'Slide' variable and show metric-specific outliers:
if (FALSE) {
savePlots(metrics,
          outdir = "./",
          colorby = "Slide",
          show_which_outliers = "metric",
          name = "example_minfidata")
}

## Change dimensions and use pdf-format
if (FALSE) {
savePlots(metrics,
          outdir = "./",
          colorby = "Slide",
          show_which_outliers = "metric",
          name = "example_minfidata",
          width = 10, height = 8, filetype="pdf"
          )
}