Update a QCmetrics object with a cleaned/normalized beta-matrix

Update a QCmetrics object with a cleaned/normalized beta-matrix

updateQCmetrics(
  qcmetrics,
  beta,
  samplesheet = NULL,
  beta_format = NULL,
  chunk_size = 500,
  read_chunks = TRUE,
  delimiter = "\t",
  calculate_PCs = FALSE,
  n_probes_PCA = NULL,
  n_estimate_PCA = chunk_size,
  nrPCs = 10,
  epidish_method = c("RPC", "CBS", "CP"),
  heatmap_variables = NULL,
  heatmap_variables_fct = NULL,
  verbose = TRUE
)

Arguments

qcmetrics

QCmetrics object

beta

beta-matrix or a filepath pointing to a txt-file where the first column contains the probe identifiers, and the remaining columns are samples.

samplesheet

samplesheet

beta_format

if a filepath is provided to the 'beta' parameter (instead of an R matrix), this parameter describes the format ('rds' or 'txt'). This is inferred from the file extension if this parameter is empty.

chunk_size

chunk size, applicable if 'read_chunks'=TRUE and a filepath to a txt-file is provided in the 'beta' parameter.

read_chunks

TRUE/FALSE, read the data in chunks? (applicable if 'beta' points toward a txt-file.)

delimiter

Defaults to `\t`, only applicable for txt-files.

calculate_PCs

should PCA be performed? (TRUE/FALSE). Default = TRUE

n_probes_PCA

number of most variable probes to maintain for PCA, defaults to 'NULL' (all probes are used).

n_estimate_PCA

number of samples used to estimate the most variable probes (defaults to 'chunk_size').

nrPCs

nr of PCs to retain. Default = 10

epidish_method

Which method to using in the EpiDish cellcount prediction (RPC, CBS, or CP)

heatmap_variables

If heatmap = y, this parameter can be used to add extra variables (in a character vector) that should be tested for an association with PCs.

heatmap_variables_fct

If extra variables are supplied for heatmap (heatmap_variables option), which of them should be treated as a factor?

verbose

Should the function be verbose? (TRUE/FALSE)